requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
aiohttp==3.8.6
pydantic==2.4.2
nltk==3.8.1